temporary folder is now always unique

+++ tests/test_xxcleanup.py
@@ -1,12 +1,30 @@
 import os
 import shutil
-from os.path import join
+from os.path import join, exists
 
 TEMP_DIR = join(os.getcwd(), "temporary")
 
 # Cleanup
-os.remove("out.txt")
-os.remove("out.bed")
-os.remove("out.log")
-# os.remove("out_altlist.txt")
-shutil.rmtree(TEMP_DIR)
+try:
+    os.remove("out.txt")
+except FileNotFoundError:
+    pass
+try:
+    os.remove("out.bed")
+except FileNotFoundError:
+    pass
+try:
+    os.remove("out.log")
+except FileNotFoundError:
+    pass
+try:
+    os.remove("ut_altlist.txt")
+except FileNotFoundError:
+    pass
+
+if exists(TEMP_DIR):
+    shutil.rmtree(TEMP_DIR)
+
+for i in range(100):
+    if exists(TEMP_DIR + str(i)):
+        shutil.rmtree(TEMP_DIR + str(i))
